% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_score.R
\name{composite_score}
\alias{composite_score}
\title{Calculate Reach or Engagement Metrics for a given Dataset}
\usage{
composite_score(data, columns, metric)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{columns}{A character vector specifying the names of the columns to be used.}

\item{metric}{A string describing the metric being estimated. Either reach or engagement.}
}
\value{
A list with two elements: a data frame with the original data and added reach metrics columns (`p`, `n` and `metric`),
and a data frame with the weights of the columns.
}
\description{
This function calculates the estimated reach or engagement metrics for a given dataset and specified columns.
It normalizes the specified columns, performs PCA, calculates the weights,
and computes the reach metrics.
}
\examples{
\dontrun{
columns_to_use <- c("page_views_total", "page_video_views", "photo_views")
result <- composite_score(facebook_reach_week, columns_to_use, "reach")
facebook_reach_week <- result$data
weights <- result$weights
}
}
