% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_engagement.R
\name{estimate_engagement}
\alias{estimate_engagement}
\title{Calculate Reach Metrics for a Dataset}
\usage{
estimate_engagement(data, columns)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{columns}{A character vector specifying the names of the columns to be used.}
}
\value{
A data frame with the original data and added reach metrics columns (`engage_p` and `engage_n`).
}
\description{
This function calculates the reach metrics for a given dataset and specified columns.
It normalizes the specified columns, performs PCA, calculates the weights,
and computes the reach metrics.
}
\examples{
\dontrun{
columns_to_use <- c("comments", "likes", "shares")
facebook_engaged <- estimate_engagement(facebook_data, columns_to_use)
}
}
