% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_reach.R
\name{estimate_reach}
\alias{estimate_reach}
\title{Calculate Reach Metrics for a Dataset}
\usage{
estimate_reach(data, columns)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{columns}{A character vector specifying the names of the columns to be used.}
}
\value{
A data frame with the original data and added reach metrics columns (`reach_p` and `reach_n`).
}
\description{
This function calculates the reach metrics for a given dataset and specified columns.
It normalizes the specified columns, performs PCA, calculates the weights,
and computes the reach metrics.
}
\examples{
\dontrun{
columns_to_use <- c("page_views_total", "page_video_views", "photo_views")
facebook_reach_week <- estimate_reach(facebook_reach_week, columns_to_use)
}
}
