% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_round_time}
\alias{fast_round_time}
\title{Round Time Values with Precision}
\usage{
fast_round_time(x, precision = 1, method = "round")
}
\arguments{
\item{x}{Numeric or POSIXct vector of time values}

\item{precision}{Numeric value specifying rounding precision. Default is 1.0}

\item{method}{Character string specifying rounding method. Options are:
\itemize{
\item "round": Round to nearest (default)
\item "floor": Round down
\item "ceiling" or "ceil": Round up
}}
}
\value{
Numeric vector of rounded values, preserving POSIXct class and
timezone attributes if input was POSIXct
}
\description{
Rounds POSIXct or numeric time values to a specified precision using
various rounding methods. It does \emph{not} work for rounding days because of timezone issues
}
\details{
For time values, precision is typically in seconds. For example:
\itemize{
\item precision = 1: Round to nearest second
\item precision = 60: Round to nearest minute
\item precision = 3600: Round to nearest hour
}
}
\examples{
\dontrun{
times <- as.POSIXct(c("2024-01-01 12:34:56", "2024-01-01 12:35:23"))
# Round to nearest minute
fast_round_time(times, precision = 60, method = "round")
}

}
