% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_csv.R
\name{fetch_csv}
\alias{fetch_csv}
\title{Read and Combine CSV Files from a S3 Folder}
\usage{
fetch_csv(bucket, folder)
}
\arguments{
\item{bucket}{A character string specifying the name of the S3 bucket.}

\item{folder}{A character string specifying the folder path within the bucket.}
}
\value{
A data.table combining all the CSV files in the specified folder.
}
\description{
This function reads all CSV files from a specified folder in a S3 bucket and combines them into one data.table.
}
\examples{
\dontrun{
combined_data <- fetch_csv("my-bucket", "my-folder")
}
}
