% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_match_values}
\alias{get_match_values}
\title{Get Matching Values from Character Vectors}
\usage{
get_match_values(strg1_list, strg2, values)
}
\arguments{
\item{strg1_list}{List of character vectors to search for}

\item{strg2}{Character vector to search in (keys)}

\item{values}{Character vector of values corresponding to strg2}
}
\value{
List of same length as strg1_list. Each element contains:
\itemize{
\item Character vector of values from 'values' corresponding to matches in strg2
\item NULL if the input list element was NULL
\item Empty character vector if no matches found
}
}
\description{
For each character vector in a list, finds matching positions in a reference
character vector and returns corresponding values.
}
\details{
Creates a lookup map from strg2 to indices, then for each query string in
strg1_list elements, retrieves all corresponding values. If a string appears
multiple times in strg2, all corresponding values are returned.
}
\examples{
\dontrun{
keys <- c("a", "b", "c", "a")
vals <- c("1", "2", "3", "4")
queries <- list(c("a", "c"), c("b"), c("d"))
get_match_values(queries, keys, vals)
# Returns: list(c("1", "4", "3"), "2", character(0))
}

}
