% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_match_values_list}
\alias{get_match_values_list}
\title{Get Matching Values from List of Character Vectors}
\usage{
get_match_values_list(strg1_list, strg2, values)
}
\arguments{
\item{strg1_list}{List of character vectors to search for (queries)}

\item{strg2}{List of character vectors to search in (keys)}

\item{values}{Character vector of values corresponding to list elements in strg2}
}
\value{
List of same length as strg1_list. Each element contains:
\itemize{
\item Character vector of values corresponding to matches
\item NULL if the input list element was NULL
\item Empty character vector if no matches found
}
}
\description{
Similar to \code{get_match_values} but searches within a list of character
vectors instead of a single character vector.
}
\details{
Creates a lookup map from all strings in strg2 list elements to their list
indices. For each query, retrieves values corresponding to the list positions
where matches were found.
}
\examples{
\dontrun{
key_list <- list(c("a", "b"), c("c", "d"), NULL, c("e"))
vals <- c("group1", "group2", "group3", "group4")
queries <- list(c("a", "c"), c("e"), c("f"))
get_match_values_list(queries, key_list, vals)
# Returns: list(c("group1", "group2"), "group4", character(0))
}

}
