% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_linear.R
\name{impute_linear}
\alias{impute_linear}
\title{Impute Missing Values}
\usage{
impute_linear(x, time = 1:length(x))
}
\arguments{
\item{x}{Numeric vector with potential missing values}

\item{time}{Numeric vector of time points. Default is 1:length(x)}
}
\value{
Numeric vector with missing values interpolated
}
\description{
Performs imputation with extrapolation on vectors with missing values.
}
\details{
Returns original vector if there are fewer than 2 non-missing values or no missing
values. Otherwise uses \code{Hmisc::approxExtrap} for interpolation/imputation with extrapolation.
}
\note{
Original function name: \code{na_int_imp}
}
\examples{
\dontrun{
x <- c(1, 2, NA, NA, 5, 6)
impute_linear(x)
}

}
