% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cumulative.R
\name{make_cumulative}
\alias{make_cumulative}
\title{Convert Demand to Cumulative Values}
\usage{
make_cumulative(demand, cumulative)
}
\arguments{
\item{demand}{Numeric vector of demand values}

\item{cumulative}{Numeric vector of cumulative values (only first element is used)}
}
\value{
Numeric vector of cumulative sum
}
\description{
Converts a demand vector to cumulative values, using the first cumulative value
as the starting point.
}
\details{
The function sets the first demand value to match the first cumulative value,
then computes the cumulative sum of the demand vector.
}
\examples{
\dontrun{
demand <- c(5, 10, 15, 20)
cumulative <- c(100, 0, 0, 0) # Only first value matters
make_cumulative(demand, cumulative)
# Returns: 100, 110, 125, 145
}

}
