% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_reach.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\value{
A numeric vector normalized to the range [0, 1].
}
\description{
This function normalizes a numeric vector to the range [0, 1].
}
\examples{
normalize(c(1, 2, 3, 4, 5))
}
