% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spms.R
\name{spms}
\alias{spms}
\title{Smooth and scale demand using SPMS method}
\usage{
spms(data, demand_col, group_col, window = 5, align = "center", offset = 0.4)
}
\arguments{
\item{data}{A \code{data.table} containing the demand and grouping columns.}

\item{demand_col}{Character, name of the column with demand values.}

\item{group_col}{Character, name of the column used for grouping (e.g., "day").}

\item{window}{Integer, size of the smoothing window (default = 5).}

\item{align}{Character, alignment for moving average: "left", "center", or "right" (default = "center").}

\item{offset}{Numeric, offset for scaling (default = 0.4).}
}
\value{
A \code{data.table} with an additional column \code{demand_spms}.
}
\description{
SPMS: Smooth – Pointwise Move – Scale was developed to mitigate the effect of
low transmission resolution of SHMs that create level-like patterns.
}
\examples{
library(data.table)
dt <- data.table(day = rep(1:2, each = 5), demand = c(10, 12, 14, 16, 18, 20, 22, 24, 26, 28))
spms(dt, demand_col = "demand", group_col = "day", window = 3, align = "center", offset = 0.4)

}
\references{
Schaffer, M., Leiria, D., Vera-Valdés, J. E., & Marszal-Pomianowska, A. (2023).
\emph{Increasing the accuracy of low-resolution commercial smart heat meter data and analysing its error.}
Proceedings of the European Conference on Computing in Construction.
DOI: \doi{10.35490/EC3.2023.208}
URL: \url{https://ec-3.org/publications/conference/paper/?id=EC32023_208}
}
