% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector}
\usage{
normalize(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector to be normalized.}

\item{na.rm}{A logical value indicating whether to remove NA values before normalization. Default is FALSE.}
}
\value{
A numeric vector normalized to the range [0, 1], maintaining the original length.
}
\description{
This function normalizes a numeric vector to the range [0, 1].
}
\examples{
normalize(c(1, 2, 3, 4, 5))
normalize(c(1, 2, NA, 4, 5), na.rm = TRUE)
}
